clean up test cases and fix api route for permission test
